move goal finished status listener for updating average health to project entry activity so when send message, don't need to navigate to progress page first
