+++ app/src/main/java/edu/northeastern/numad23sp_team16/Project/ProjectEntryActivity.java
@@ -1,5 +1,7 @@
 package edu.northeastern.numad23sp_team16.Project;
 
+import static android.content.ContentValues.TAG;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
@@ -27,6 +29,13 @@ import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
 
 import java.sql.Timestamp;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.Calendar;
+import java.util.Date;
+import java.util.Locale;
+import java.util.Objects;
+import java.util.concurrent.TimeUnit;
 
 import edu.northeastern.numad23sp_team16.R;
 import edu.northeastern.numad23sp_team16.models.Message;
@@ -39,6 +48,14 @@ public class ProjectEntryActivity extends AppCompatActivity {
     private final int PERMISSION_REQUEST_CODE = 0;
     private DatabaseReference messagesRef;
     private ChildEventListener messagesChildEventListener;
+
+    private SimpleDateFormat dateFormat = new SimpleDateFormat("MM/dd/yy", Locale.US);
+    private DatabaseReference mDatabase;
+    private DatabaseReference petHealthRef;
+    private float totalHealth;
+    private int totalDays = 1;
+    private ValueEventListener goalFinishedStatusPostListener;
+
 
     private static final String CURRENT_USER = "CURRENT_USER";
     private final String LOGIN_TIME = "LOGIN_TIME";
@@ -60,6 +77,103 @@ public class ProjectEntryActivity extends AppCompatActivity {
         }
         Log.i("ProjectEntry", "currentUser from bundle: " + currentUser);
         Log.i("ProjectEntry", "loginTime from bundle: " + loginTime);
+
+
+
+        // Connect to firebase database
+        mDatabase = FirebaseDatabase.getInstance().getReference("FinalProject");
+
+        // Create reference to GoalFinishedStatus node in database
+        DatabaseReference goalFinishedStatusRef = mDatabase.child("GoalFinishedStatus");
+
+        // Get user's pet's health node from database and create listener
+        petHealthRef = mDatabase.child("PetHealth")
+                .child("health" + currentUser);
+
+        // Get today's date with time of 0
+        Calendar calendar = Calendar.getInstance();
+        calendar.set(Calendar.MILLISECOND, 0);
+        calendar.set(Calendar.SECOND, 0);
+        calendar.set(Calendar.MINUTE, 0);
+        calendar.set(Calendar.HOUR, 0);
+        Date currentDate = calendar.getTime();
+
+        // Calculate and assign number of days it has been between current date and creation date
+        petHealthRef.addListenerForSingleValueEvent(new ValueEventListener() {
+            @Override
+            public void onDataChange(DataSnapshot dataSnapshot) {
+                String creationDate = dataSnapshot.child("creationDate").getValue(String.class);
+                totalDays = calculateNumberOfDays(creationDate);
+            }
+
+            @Override
+            public void onCancelled(DatabaseError databaseError) {
+                Log.d(TAG, "onCancelled: Database error retrieving creation date");
+            }
+        });
+
+        // Create listener for changes to GoalFinishedStatus
+        goalFinishedStatusPostListener = new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot snapshot) {
+                totalHealth = 0;
+
+                // Iterate through GoalFinishedStatus nodes
+                for (DataSnapshot data : snapshot.getChildren()) {
+                    String user = data.child("userId").getValue(String.class);
+
+                    // Check if goal finished status is associated with current user
+                    if (Objects.equals(user, currentUser)) {
+                        Log.d(TAG, "onDataChange: user " + user);
+                        Log.d(TAG, "onDataChange: user current " + currentUser);
+
+                        // Get the date
+                        DataSnapshot dateMap = data.child("dateMap");
+                        int year = dateMap.child("year").getValue(Integer.class);
+                        int month = dateMap.child("month").getValue(Integer.class);
+                        int day = dateMap.child("day").getValue(Integer.class);
+
+                        // Convert stored date to Date object with time of 0
+                        Calendar calendar = Calendar.getInstance();
+                        calendar.set(year, month - 1, day);
+                        calendar.set(Calendar.MILLISECOND, 0);
+                        calendar.set(Calendar.SECOND, 0);
+                        calendar.set(Calendar.MINUTE, 0);
+                        calendar.set(Calendar.HOUR, 0);
+                        Date date = calendar.getTime();
+
+                        Log.d(TAG, "onDataChange: current date " + currentDate);
+                        Log.d(TAG, "onDataChange: string date " + date);
+
+                        // Only calculate into pet's health if not the current day
+                        if (!currentDate.equals(date)) {
+
+                            // Add to total health
+                            totalHealth += data.child("percentageOfToday").getValue(Float.class);
+                            Log.d(TAG, "onDataChange: total health " + totalHealth);
+
+                            // Calculate average health from total health and number of days
+                            float averageHealth = totalHealth / totalDays;
+                            Log.d(TAG, "onDataChange: total days " + totalDays);
+
+                            // Update average health for PetHealth node
+                            petHealthRef.child("averageHealth").setValue(averageHealth);
+                            Log.d(TAG, "onDataChange: averageHealth " + averageHealth);
+                        }
+                    }
+                }
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError error) {
+                // Getting GoalFinishedStatus failed, log a message
+                Log.w(TAG, "Error getting goal finished status from database");
+            }
+        };
+        goalFinishedStatusRef.addValueEventListener(goalFinishedStatusPostListener);
+
+
+
 
         // TODO: change the hardcoded heartCount to user's pet heartCount from database
         int heartCount = 8;
@@ -114,6 +228,34 @@ public class ProjectEntryActivity extends AppCompatActivity {
 
         messagesRef.addChildEventListener(messagesChildEventListener);
     }
+
+
+
+    // Calculate total number of days between creation date and current day
+    private int calculateNumberOfDays(String date) {
+        // Convert creation date to Date object
+        Date creationDate = null;
+        try {
+            creationDate = dateFormat.parse(date);
+        } catch (ParseException e) {
+            Log.d(TAG, "calculateNumberOfDays: Error parsing date");
+        }
+
+        Date currentDate = new Date();
+
+        // Calculate duration between the current date and creation date
+        long durationInMillis = currentDate.getTime() - creationDate.getTime();
+        int differenceInDays = (int) TimeUnit.DAYS.convert(durationInMillis, TimeUnit.MILLISECONDS);
+        Log.d(TAG, "calculateNumberOfDays: " + differenceInDays);
+
+        if (differenceInDays == 0) {
+            return 1;
+        }
+        return differenceInDays;
+    }
+
+
+
 
     public void startProfileActivity(View view) {
         // Pass the current user info to Profile activity -- Yutong
